feat(ledger): add ledger desktop support
