+++ desktop/wallet/package.json
@@ -45,6 +45,7 @@
     "electron": "^5.0.2",
     "electron-builder": "^20.41.0",
     "jest": "^24.8.0",
+    "os": "^0.1.1",
     "rimraf": "^2.6.3"
   },
   "update": {
@@ -55,11 +56,14 @@
   },
   "dependencies": {
     "@burstjs/http": "^0.1.2",
+    "@ledgerhq/hw-transport": "^4.63.2",
+    "@ledgerhq/hw-transport-node-hid": "^4.63.2",
     "electron-dl": "^1.14.0",
     "get-ssl-certificate": "^2.3.2",
     "js-yaml": "^3.13.1",
     "lodash": "^4.17.11",
     "prop-types": "^15.7.2",
+    "regenerator-runtime": "^0.13.2",
     "semver": "^6.1.1",
     "snyk": "^1.170.0",
     "winston": "^3.2.1"
