--- scripts/noop
@@ -1,1 +1,0 @@
-export const NOOP = 0;
